+++ compiler/maat_eval/src/lib.rs
@@ -1,0 +1,12 @@
+mod builtins;
+mod env;
+mod eval;
+mod macros;
+mod object;
+
+pub use env::Env;
+pub use eval::eval;
+pub use macros::{define_macros, expand_macros};
+pub use object::{
+    BuiltinFn, FALSE, Function, HashObject, Hashable, Macro, NULL, Object, Quote, TRUE,
+};